Heretic: add support for extended nodes
Applied: https://github.com/fabiangreffrath/crispy-doom/commit/d3b76990baecb0cd7390d21ba30bf1243c442521

+++ src/heretic/r_local.h
@@ -117,9 +117,10 @@ typedef struct line_s
 {
     vertex_t *v1, *v2;
     fixed_t dx, dy;             // v2 - v1 for side checking
-    short flags;
+    unsigned short flags;
     short special, tag;
-    short sidenum[2];           // sidenum[1] will be -1 if one sided
+    // sidenum[1] will be -1 (NO_INDEX) if one sided
+    unsigned short sidenum[2];
     fixed_t bbox[4];
     slopetype_t slopetype;      // to aid move clipping
     sector_t *frontsector, *backsector;
@@ -131,8 +132,8 @@ typedef struct line_s
 typedef struct subsector_s
 {
     sector_t *sector;
-    short numlines;
-    short firstline;
+    int numlines;
+    int firstline;
 } subsector_t;
 
 typedef struct
@@ -152,7 +153,7 @@ typedef struct
 {
     fixed_t x, y, dx, dy;       // partition line
     fixed_t bbox[2][4];         // bounding box for each child
-    unsigned short children[2]; // if NF_SUBSECTOR its a subsector
+    int children[2];            // if NF_SUBSECTOR its a subsector
 } node_t;
